requirements.txt
pygame-ce==2.5.5
pytest==8.4.0

